move timer.hpp to common
