+++ power_grid_model_c/power_grid_model/include/power_grid_model/math_solver/iterative_linear_se_solver.hpp
@@ -18,7 +18,7 @@
 #include "../common/exception.hpp"
 #include "../common/power_grid_model.hpp"
 #include "../common/three_phase_tensor.hpp"
-#include "../timer.hpp"
+#include "../common/timer.hpp"
 
 namespace power_grid_model::math_solver {
